requirements.txt
Flask~=3.1.0
getmac~=0.9.5
winshell~=0.6
pyinstaller
pycryptodome~=3.21.0
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
Werkzeug~=3.1.3
wmi
python-dotenv